% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/building_area.R
\name{building_area}
\alias{building_area}
\title{Calculate building area}
\usage{
street_width (building, street, k = ，w =  , option = c("max","near","point"))
}
\arguments{
\item{option}{Defines the method by which building area is calculated.
If option = "fixed", k should be defined by the user (default = 40 meters).
If option = "street", k equals to the width of street on each side.}

\item{k}{Defines the distance from streets. Buildings within which area will be considered as buildings alongside streets.}

\item{building}{A Simple Feature containing the building footprints of a city}

\item{street}{A Simple Feature containing the street network of a city.}
}
\description{
Calculates the area of buildings on two sides of streets. It performs geometric translation to the lines ("street") by distance k, and defines polygons ("building") that intersect with the lines as building alongside streets.
}
\details{
Calculate building area with two options available.
It returns the summary statistics (max, min, average, weighted average, standard deviation) of building area alongside every street to the Simple Feature "street". If the calculation is performed in option "street", the width of street should first be calculated
If the translation lines of streets and the polygons ("building") does not have intersection, it will return NA.
}
\examples{
## Not run:
building <-data(building)
street<-data(street)
building_area (building, street, option = "fixed", k = 40)
building <-data(building)
street<-data(street)
building_area (building, street, option = "street", k = 40)
## End(Not run)
}
