% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/road_continuity.R
\name{continuity}
\alias{continuity}
\title{Calculate the continuity of street buildings}
\usage{
building_continuity(building, street, di, option = ("regular", "near"))
}
\arguments{
\item{di}{Defines the distance from streets. Buildings within which area will be considered as buildings alongside street (default = 40 meters).}

\item{option}{option     defines the method by which continuity of street buildings is calculated.
If option = "regular", the projection length of buildings is defined as the length of all building projection on each side of streets.
If option = "near", the projection length of buildings is defined as the length of projection of the nearest building edge to the street segment.}

\item{building}{A Simple Feature containing the building footprints of a city}

\item{street}{A Simple Feature containing the street network of a city.}

\item{w}{defines the total distance of the translation of the street segment (line) in option1/2 (default = 40 meters). Defines the length of perpendicular line in option3 (default = 100 meters).}
}
\description{
It generates a buffer area with distance di on each side of street. It then projects buildings within this area onto street lines. Continuity of buildings equals the ratio of projection length of buildings to the length of streets.
In option regular, the projection length of buildings is defined as the length of all building projection on each side of streets.
In option near, the projection length of buildings is defined as the length of projection of one side of building which has the shortest distance to streets.
}
\details{
It returns the continuity (0 to 1) of the buildings alongside every street to the simple feature "street". If the translation lines of streets and the polygons ("city") does not have intersection, it will return NA.
}
\examples{
## Not run:
## Not run:
building <-data(building)
street<-data(street)
continuity(building, street, di = 40, option = "regular")
building <-data(building)
street<-data(street)
continuity(building, street, di = 40, option = "near")
## End(Not run)
}
