% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/road_width.R
\name{street_width}
\alias{street_width}
\title{Calculate the width of street defined by buildings}
\usage{
street_width (building, street, k = ，w =  , option = c("max","near","point"))
}
\arguments{
\item{k}{Defines the distance of every translation of the street segments (line) in option1/2 (default = 5 meters). Defines the distance between sample points in option3 (default = 200 meters).}

\item{w}{defines the total distance of the translation of the street segment (line) in option1/2 (default = 40 meters). Defines the length of perpendicular line in option3 (default = 100 meters).}

\item{option}{defines the method by which building-to-building width is calculated.
If option = "max", building-to-building width is calculated as the distance between translation lines of street which have the max intersection with buildings on each side and the original street line.
If option = "near", building-to-building width is calculated as the minimal distance between street and buildings on each side of streets.
If option = "point", building-to-building width is calculated as the average length of the perpendicular lines that intersects with each side of street.}

\item{building}{A Simple Feature containing the building footprints of a city}

\item{street}{A Simple Feature containing the street network of a city.}
}
\description{
Calculates  the width of the street space defined by the buildings on the two sides of the street with three options available.
In option max and 2 near, it performs geometric translation to the spatial lines ("street") by distance w (by translating lines with distance k each time).
In option max, building-to-building width is defined as the distance between original street lines and translation lines of street which have the max intersection with buildings. In option near, building-to-building width is defined as the distance between translation lines and street when translation lines first intersect with buildings.
In option point, it generates sample points by distance k alongside street and make perpendicular lines with length w. Building-to-building width is defined as the average length of the perpendicular lines that intersects with each side of streets.
}
\details{
It returns the distance from the street to the buildings on its left and right sides to the simple feature "street". If the translation lines ("street") and the polygons ("building") does not have intersection, it will return NA.
}
\examples{
## Not run:
building <-data(building)
street<-data(street)
street_width(building, street, k = 10, w = 50, option = "near")
}
